@@ -1,61 +1,52 @@
 {
-  "name": "fe",
+  "name": "frontend",
   "version": "0.0.0",
   "scripts": {
-    "start": "node --max_old_space_size=8048 ./node_modules/@angular/cli/bin/ng serve --aot --watch --proxy-config proxy.conf.json",
-    "build": "node --max_old_space_size=8048 ./node_modules/@angular/cli/bin/ng build --aot --prod"
+    "ng": "ng",
+    "start": "node --max_old_space_size=8048 ./node_modules/@angular/cli/bin/ng serve --watch --proxy-config proxy.conf.json",
+    "build": "node --max_old_space_size=8048 ./node_modules/@angular/cli/bin/ng build --prod",
+    "test": "ng test",
+    "lint": "ng lint",
+    "e2e": "ng e2e",
+    "build:prod": "NODE_ENV=production ng build --prod"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~9.0.1",
-    "@angular/cdk": "^9.0.0",
-    "@angular/common": "~9.0.1",
-    "@angular/compiler": "~9.0.1",
-    "@angular/core": "~9.0.1",
-    "@angular/forms": "~9.0.1",
-    "@angular/material": "^9.0.0",
-    "@angular/platform-browser": "~9.0.1",
-    "@angular/platform-browser-dynamic": "~9.0.1",
-    "@angular/router": "~9.0.1",
-    "@crystalui/angular-lightbox": "^1.1.8",
-    "@ctrl/ngx-emoji-mart": "^3.1.0",
-    "@eventi/interfaces": "file:../interfaces",
-    "@ivylab/overlay-angular": "0.0.6",
-    "angular-responsive-carousel": "^1.2.9",
-    "cronstrue": "^1.94.0",
-    "highcharts-angular": "^2.7.0",
-    "moment": "^2.24.0",
-    "ng-click-outside": "^6.0.0",
-    "ng-dynamic-component": "^6.1.0",
-    "ng2-tooltip-directive": "^2.9.22",
-    "ngx-cookie-service": "^3.0.3",
-    "ngx-moment": "^3.5.0",
-    "ngx-widget-grid": "^9.0.0",
-    "rxjs": "~6.5.4",
-    "tslib": "^1.10.0",
+    "@angular/animations": "~11.0.6",
+    "@angular/cdk": "^11.0.3",
+    "@angular/common": "~11.0.6",
+    "@angular/compiler": "~11.0.6",
+    "@angular/core": "~11.0.6",
+    "@angular/forms": "~11.0.6",
+    "@angular/material": "^11.0.3",
+    "@angular/platform-browser": "~11.0.6",
+    "@angular/platform-browser-dynamic": "~11.0.6",
+    "@angular/router": "~11.0.6",
+    "@eventi/interfaces": "../interfaces",
+    "ng-click-outside": "^7.0.1",
+    "ngx-cookie-service": "^11.0.2",
+    "rxjs": "~6.6.0",
+    "tslib": "^2.0.0",
     "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.900.2",
-    "@angular/cli": "~9.0.2",
-    "@angular/compiler-cli": "~9.0.1",
-    "@angular/language-service": "~9.0.1",
-    "@types/gridstack": "^0.5.2",
-    "@types/highcharts": "^7.0.0",
-    "@types/jasmine": "~3.5.0",
-    "@types/jasminewd2": "~2.0.3",
+    "@angular-builders/custom-webpack": "^11.0.0",
+    "@angular/cli": "~11.0.6",
+    "@angular/compiler-cli": "~11.0.6",
+    "@types/jasmine": "~3.6.0",
     "@types/node": "^12.11.1",
-    "codelyzer": "^5.1.2",
-    "jasmine-core": "~3.5.0",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.3.0",
-    "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage-istanbul-reporter": "~2.1.0",
-    "karma-jasmine": "~2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.2",
-    "protractor": "~5.4.3",
+    "autoprefixer": "^10.0.2",
+    "codelyzer": "^6.0.0",
+    "dotenv": "^8.2.0",
+    "ngx-build-plus": "^10.1.1",
+    "ngx-tailwind": "^1.1.0",
+    "postcss": "^8.1.7",
+    "postcss-import": "^13.0.0",
+    "postcss-loader": "^4.0.4",
+    "postcss-scss": "^3.0.4",
+    "tailwindcss": "^2.0.1",
     "ts-node": "~8.3.0",
-    "tslint": "~5.18.0",
-    "typescript": "~3.7.5"
+    "tslint": "~6.1.0",
+    "typescript": "~4.0.2"
   }
 }
