@@ -1,8 +1,11 @@
 import { ICacheable } from "../app.interfaces";
-import { IErrorResponse, IFormErrorField } from "@eventi/interfaces";
+import { IErrorResponse, IFormErrorField, Y } from "@eventi/interfaces";
 import { FormGroup } from "@angular/forms";
 
-export const handleFormErrors = (obj: ICacheable<any>, error: IErrorResponse): ICacheable<any> => {
+export const handleFormErrors = (
+  obj: ICacheable<any>,
+  error: IErrorResponse
+): ICacheable<any> => {
   // Objects pass by reference & reset fields, stay pure please :)
   const o = Object.assign({}, obj);
   o.error = null;
@@ -13,30 +16,46 @@ export const handleFormErrors = (obj: ICacheable<any>, error: IErrorResponse): I
     {}
   );
 
-  // Put general error from message
+  // Put the form error message in the Cachable
   o.error = error.message;
 
-  // Map error to each field from response to cacheable form error field
-  Object.values(error.errors).forEach((error: IFormErrorField) => {
-    // if (o.form_errors.hasOwnProperty(error.param)) {
-    //   o.form_errors[error.param] = error.msg;
-    // }
-  });
+  // Recursively map the error response to a KV param:ErrCode object
+  o.form_errors = Y<IFormErrorField[][], typeof o.form_errors>((r) => (f) => {
+    return f.reduce((acc, curr) => {
+      if (curr.nestedErrors?.length > 0) {
+        acc[curr.param] = r(curr.nestedErrors);
+      } else {
+        acc[curr.param] = curr.code;
+      }
+      return acc;
+    }, {});
+  })(error.errors);
 
   return o;
 };
 
-export const displayValidationErrors = (formGroup: FormGroup, cacheable: ICacheable<any>) => {
-  Object.keys(formGroup.controls).forEach((field) => {
-    const control = formGroup.get(field);
-    if (cacheable.form_errors[field]) {
-      control.setErrors({ backendIssue: cacheable.form_errors[field] });
-    }
-  });
+export const displayValidationErrors = (
+  formGroup: FormGroup,
+  cacheable: ICacheable<any>
+) => {
+  // Recursively go through the controls setting the error messages
+  Y<[FormGroup, string], void>(r => (fg, path) => {
+    Object.keys(fg.controls).forEach((f) => {
+      const control = fg.get(f);
+      const i = path ? path + `.${f}` : f; // make a.b.c.d for object path
+      // Get error by path, if one exists in the cacheable form errors
+      const e = i.split(".").reduce((acc, val) => acc && acc[val], cacheable.form_errors);
+      if(e && typeof(e) == "string") {
+        control.setErrors({ backendIssue: e });
+        control.markAsTouched(); // trigger change detection
+      };
+      if((control as FormGroup).controls) r(control as FormGroup, i);
+    });
+  })(formGroup, null);
 };
 
-export const formHasValidationErrors = (formGroup:FormGroup):boolean => {
-  return Object.keys(formGroup.contains).some(field => {
+export const formHasValidationErrors = (formGroup: FormGroup): boolean => {
+  return Object.keys(formGroup.contains).some((field) => {
     return formGroup.get(field).invalid;
-  })
-}
+  });
+};
