@@ -16,11 +16,13 @@ describe("Verify the Onboarding process pulls in the host relationship", async (
         client = await Stories.actions.users.createUser(UserType.Client);
         await Stories.actions.common.switchActor(UserType.Client)
         await Stories.actions.common.setup();
+        
         host = await Stories.actions.hosts.createHost({
             username: 'somecoolhost',
             name: 'Some Cool Host',
             email_address: 'host@cass.si',
         });
+
         let onboarding = await Stories.actions.hosts.readOnboardingProcessStatus(host);
         expect(onboarding.host).to.exist;
         expect(onboarding.created_at).to.exist;
@@ -28,10 +30,7 @@ describe("Verify the Onboarding process pulls in the host relationship", async (
         expect(onboarding.last_modified_by._id).to.eq(client._id);
         expect(onboarding.last_modified_by.name).to.eq(client.name);
         expect(onboarding.last_modified_by.username).to.eq(client.username);
-        
-           
     }); 
-    
 });
     
     
