@@ -1,39 +1,31 @@
-// import Axios from 'axios';
-// import { describe, it } from 'mocha';
-// import { expect } from 'chai';
-// import { IUser } from '@eventi/interfaces';
-// import { Stories } from '../../stories';
+import Axios from 'axios';
+import { describe, it } from 'mocha';
+import { expect } from 'chai';
+import { IUser } from '@eventi/interfaces';
+import { Stories } from '../../stories';
+import { UserType } from '../../environment';
 
-// describe('As a user, I want to be able to CRUD', async () => {
-//   let user: IUser;
+describe('As a user, I want to be able to CRUD', async () => {
+  let user: IUser;
 
-//   it('Should create a user', async () => {
-//     user = await Stories.actions.users.createUser({
-//       username: 'cass',
-//       email_address: 'm@cass.si',
-//       password: 'helloworld',
-//     });
+  it('Should create a user', async () => {
+    user = await Stories.actions.users.createUser(UserType.Member);
 
-//     expect(user).to.not.be.null;
-//     expect(user.name).to.be.eq('cass');
-//     expect(user.email_address).to.be.eq('m@cass.si');
-//   });
+    expect(user).to.not.be.null;
+    expect(user.name).to.be.eq('cass');
+  });
 
-//   it('Should get the newly created user', async () => {});
+  it('Should get the newly created user', async () => {});
 
-//   it('Should update a user & ensure only certain fields can be modified', async () => {});
+  it('Should update a user & ensure only certain fields can be modified', async () => {});
 
-//   it('Should delete a user', async () => {});
-// });
+  it('Should delete a user', async () => {});
+});
 
-// describe('Error checking for user CRUD', async () => {
-//   let user: IUser;
+describe('Error checking for user CRUD', async () => {
+  let user: IUser;
 
-//   it('Should not allow me to create an account with a pre-existing username/email-address', async () => {
-//     user = await Stories.actions.users.createUser({
-//       username: 'cass',
-//       email_address: 'm@cass.si',
-//       password: 'helloworld',
-//     });
-//   });
-// });
+  it('Should not allow me to create an account with a pre-existing username/email-address', async () => {
+    user = await Stories.actions.users.createUser(UserType.Member, true);
+  });
+});
